blogapp/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-ckeditor==6.1.0
django-js-asset==2.0.0
Pillow==8.3.2
pytz==2020.1
sqlparse==0.4.3
pysqlite3
pysqlite3-binary